video-highlight-extractor/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-dotenv==1.0.1
moviepy==1.0.3
google-generativeai==0.4.0
pgvector==0.2.3
SpeechRecognition==3.10.0
pydub==0.25.1
tqdm==4.66.1